moved ContentCard to organisms
